# packages/contracts/openapi/alo-noon.v1.yaml
openapi: 3.1.0
info:
  title: Alo Noon API
  version: 0.11.0
  description: >-
    Phase 2E contract. Operational endpoints, tenant-aware discovery, identity,
    customer-owned addresses, authoritative delivery quotes, and atomic
    capacity-backed Quote-to-Order conversion, and the internal payment and
    double-entry ledger foundation are implemented. Payment execution remains
    internal; the versioned tenant chart of accounts and governed account state
    foundation are implemented. The tenant-scoped, SPI-versioned payment-provider abstraction,
    opaque credential-reference, payment-attempt, and callback replay foundation
    are implemented without a real gateway adapter. A customer may open the payment for an
    order they own and read it back, and nothing more: no public endpoint lets a client name
    a payment amount or assert a payment state, and no ledger-governance write endpoint is
    exposed at all. The authenticated initialization orchestrator is implemented as an
    injectable, fail-closed boundary. Gateway callbacks are recorded and then settled by
    server-to-server verification: a payment is captured, posted to the double-entry ledger,
    and marked paid on its order only when the gateway confirms it for the exact expected
    amount. No settlement surface is exposed publicly — the customer's return carries an
    opaque reference and no verdict. Authentication delivery
...
servers:
  - url: http://localhost:3001
paths:
...
  /api/v1/payments:
    post:
      operationId: startPayment
      summary: Open the payment for an order the caller owns.
      description: >-
        The amount is never sent — it comes from the order's own total, because a
        client-supplied amount would be a client-chosen price. The payment is opened in
        CREATED and stays there: settlement walks it to PENDING, AUTHORIZED and captured
        when the gateway confirms, and letting a customer request any of those would be
        letting them assert they had paid. Replaying the same idempotency key returns the
        same payment rather than opening a second.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentCheckoutStart' }
      responses:
        '201':
          description: The opened payment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentEnvelope' }
        '400': { description: The request is malformed. }
        '401': { description: A valid customer session is required. }
        '404': { description: No such order for this customer. }
        '409': { description: This order already has a payment. }
        '422': { description: The order cannot be paid in its current state. }
        '503': { description: Payments are temporarily unavailable. }
  /api/v1/payments/{paymentId}:
    get:
      operationId: readPayment
      summary: Read a payment back, after returning from the gateway.
      description: >-
        The screen a gateway returns to needs this. The return redirect proves nothing on
        its own — every parameter on it is attacker-controllable, and the verdict comes
        from the gateway's own server-to-server answer — so the app asks what actually
        happened rather than believing the URL it landed on. Settlement may not have
        finished when the customer returns, so a state short of CAPTURED is a reason to
        ask again, not a failure.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The payment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentEnvelope' }
        '401': { description: A valid customer session is required. }
        '404': { description: No such payment for this customer. }
        '503': { description: Payments are temporarily unavailable. }
...
  /api/v1/admin/withdrawals:
    get:
      operationId: listOpenWalletWithdrawals
      summary: Customer withdrawal requests still awaiting an answer, oldest first.
      description: >-
        The amount has already left each customer's balance — it is debited when the
        request is made, not when the transfer is sent, so nothing here is spendable while
        somebody at a bank is already sending it. Only the last four digits of the card
        are stored, which with the holder's name is what a bank transfer form needs.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Open requests, with the customer who made each.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaffWithdrawalListEnvelope' }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle customer withdrawals. }
        '503': { description: The withdrawal desk is temporarily unavailable. }
  /api/v1/admin/withdrawals/{withdrawalId}/paid:
    post:
      operationId: markWalletWithdrawalPaid
      summary: Record that a customer's withdrawal was sent.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BankReferenceCommand' }
      responses:
        '200':
          description: The withdrawal, now settled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletWithdrawalEnvelope' }
        '400': { description: The bank reference is missing or too long. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle customer withdrawals. }
        '404': { description: No such withdrawal request. }
        '503': { description: The withdrawal desk is temporarily unavailable. }
  /api/v1/admin/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWalletWithdrawal
      summary: Refuse a withdrawal and put the money back.
      description: >-
        The credit lands in the same transaction as the refusal, as its own statement
        line, so the customer sees what happened rather than a balance that silently
        returned. The reason is required and the customer reads it verbatim — a rejection
        with no reason is a phone call they have to make to find out anything.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectWithdrawalCommand' }
      responses:
        '200':
          description: The withdrawal, refused, with the money back on the balance.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletWithdrawalEnvelope' }
        '400': { description: The reason is missing or too long. }
        '401': { description: A valid staff session is required. }
        '403': { description: The account may not settle customer withdrawals. }
        '404': { description: No such withdrawal request. }
        '503': { description: The withdrawal desk is temporarily unavailable. }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: alo_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp, version]
      properties:
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        version: { type: string, const: v1 }
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^\d+$' }
        currency: { type: string, enum: [IRR] }
...
    PaymentAggregateState:
      type: string
      enum: [CREATED, PENDING, AUTHORIZED, CAPTURED, FAILED]
...
    PaymentCheckoutStart:
      type: object
      additionalProperties: false
      required: [orderId, idempotencyKey]
      properties:
        orderId: { type: string, format: uuid }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
    PaymentEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/PaymentSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
    PaymentSummary:
      type: object
      required: [id, publicId, orderId, customerId, state, amount, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        publicId: { type: string, minLength: 8, maxLength: 32 }
        orderId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        state: { $ref: '#/components/schemas/PaymentAggregateState' }
        amount: { $ref: '#/components/schemas/Money' }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    BankReferenceCommand:
      type: object
      additionalProperties: false
      required: [bankReference]
      properties:
        bankReference:
          type: string
          minLength: 1
          maxLength: 128
          description: >-
            The bank's own reference. Required, because a transfer recorded with nothing
            to look it up by cannot be reconciled against a statement, and reconciling
            against a statement is the entire point of recording it.
    RejectWithdrawalCommand:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 500
          description: >-
            The customer reads this verbatim. A rejection with no reason is a phone call
            they have to make to find out anything, and their money is already back.
    WalletWithdrawalSummary:
      type: object
      required: [id, amount, state, cardLastFour, cardHolderName, requestedAt]
      properties:
        id: { type: string, format: uuid }
        amount: { $ref: '#/components/schemas/Money' }
        state: { type: string, enum: [REQUESTED, PAID, REJECTED] }
        cardLastFour:
          type: string
          pattern: '^[0-9]{4}$'
          description: >-
            Four digits and no more. The customer sends a full card number once; the API
            keeps these and forgets the rest before the row is written.
        cardHolderName: { type: string, minLength: 1, maxLength: 120 }
        iban: { type: string, pattern: '^IR[0-9]{24}$' }
        bankReference: { type: string, minLength: 1, maxLength: 128 }
        rejectionReason: { type: string, minLength: 1, maxLength: 500 }
        requestedAt: { type: string, format: date-time }
        settledAt: { type: string, format: date-time }
    WalletWithdrawalEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: { $ref: '#/components/schemas/WalletWithdrawalSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    StaffWithdrawalSummary:
      allOf:
        - $ref: '#/components/schemas/WalletWithdrawalSummary'
        - type: object
          required: [customerId, customerMobileE164]
          properties:
            customerId: { type: string, format: uuid }
            customerMobileE164: { type: string, pattern: '^\+989\d{9}$' }
    StaffWithdrawalListEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/StaffWithdrawalSummary' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...